fix web3 modal
