--- src/modules/modal/index.js
@@ -1,1 +1,0 @@
-export {default as Modal} from './modal';
