+++ src/modules/web3/index.js
@@ -1,0 +1,1 @@
+export {default as Web3} from './modal';
